fix(#7407): satisfy qulice for the snapshot of empty attributes
The freeze moves out of the constructor, which must stay free of method
calls, into `PhDefault.copy()`. The two maps of attributes walk their
pairs through the new `Walkable`, so the dispatch reads an interface
and not an implementation type. The two regressions move into
`CopiedAttrsTest`, where they belong and where they do not push
`PhDefaultTest` past the file length cap.
