+++ eo-runtime/src/test/java/org/eolang/CopiedAttrsTest.java
@@ -22,6 +22,35 @@ import org.junit.jupiter.api.Test;
 final class CopiedAttrsTest {
 
     @Test
+    void keepsCopyEmptyAfterLatePutOnOrigin() {
+        final Phi origin = new PhDefault(new Attrs(new Attr("v", new AtVoid("v"))));
+        final Phi copy = origin.copy();
+        origin.put("v", new Data.ToPhi(7L));
+        Assertions.assertThrows(
+            ExAbstract.class,
+            () -> new Dataized(new PhSafe(copy).take("v")).take(),
+            "a copy must not carry a value the origin received after the copy was made"
+        );
+    }
+
+    @Test
+    void readsTheSameCopyWhicheverWayItIsOrdered() {
+        final Phi origin = new PhDefault(new Attrs(new Attr("v", new AtVoid("v"))));
+        final Phi copy = origin.copy();
+        Assertions.assertThrows(
+            ExAbstract.class,
+            () -> new Dataized(new PhSafe(copy).take("v")).take(),
+            "a copy of an object with an empty attribute must start empty, but it didnt"
+        );
+        origin.put("v", new Data.ToPhi(7L));
+        Assertions.assertThrows(
+            ExAbstract.class,
+            () -> new Dataized(new PhSafe(copy).take("v")).take(),
+            "reading a copy before the origin was filled must not change what the copy holds"
+        );
+    }
+
+    @Test
     void copiesNoAttributeThatNobodyTakes() {
         final AtomicInteger copies = new AtomicInteger();
         new PhDefault(
